Use v_flop + move (test) ngcwriter to the attic
